+++ src/core/request/response/index.ts
@@ -15,9 +15,6 @@ import { EventEmitter2 as EventEmitter } from 'eventemitter2';
 import { once, deprecated } from 'core/functools';
 import { IS_NODE } from 'core/env';
 
-import { readonly } from 'core/object/proxy-readonly';
-import { clone } from 'core/object/proxy-clone';
-
 import { convertIfDate } from 'core/json';
 import { getDataType } from 'core/mime-type';
 
@@ -657,10 +654,10 @@ export default class Response<
 
 				Object.defineProperty(data, 'valueOf', {
 					configurable: true,
-					value: () => clone(originalData)
+					value: () => Object.fastClone(originalData, {freezable: false})
 				});
 
-				data = readonly(data);
+				data = Object.freeze(data);
 			}
 
 			return data;